EPMLSTR-435 refactoring
