+++ src/main/java/com/epam/test_generator/controllers/PasswordResetController.java
@@ -21,7 +21,6 @@ public class PasswordResetController {
 
     @RequestMapping(value = "/passwordReset", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
     public ResponseEntity passwordReset(@RequestBody @Valid PasswordResetDTO passwordResetDTO) {
-
         passwordService.passwordReset(passwordResetDTO);
 
         return new ResponseEntity(HttpStatus.OK);
@@ -30,7 +29,6 @@ public class PasswordResetController {
 
     @GetMapping("/passwordReset")
     public ResponseEntity displayResetPasswordPage(@RequestParam String token) {
-
         tokenService.checkToken(token);
 
         return new ResponseEntity<>(token, HttpStatus.OK);
