+++ src/main/java/com/epam/test_generator/services/EmailService.java
@@ -19,10 +19,7 @@ import javax.servlet.http.HttpServletRequest;
 @Component
 public class EmailService {
     @Autowired
-    private TokenService tokenService;
-
-    @Autowired
-    PasswordService passwordService;
+    private PasswordService passwordService;
 
     @Autowired
     private JavaMailSender emailSender;
@@ -30,19 +27,14 @@ public class EmailService {
     @Autowired
     private JavaMailSenderImpl javaMailSender;
 
+    @Autowired
+    private TokenService tokenService;
+
     @Resource
     private Environment environment;
 
-    public void sendSimpleMessage(String to, String subject, String text) {
-        SimpleMailMessage message = new SimpleMailMessage();
-        message.setTo(to);
-        message.setSubject(subject);
-        message.setText(text);
-        emailSender.send(message);
-    }
-
     public void sendRegistrationMessage(User user, HttpServletRequest request) {
-        Token userConformationToken = tokenService.createToken(user,1440);
+        Token userConformationToken = tokenService.createToken(user, 1440);
         String confirmUrl = passwordService.createConfirmUrl(request, userConformationToken);
         String subject = environment.getProperty("subject.registration.message");
         String text = environment.getProperty("registration.message");
@@ -52,14 +44,20 @@ public class EmailService {
     }
 
     public void sendResetPasswordMessage(User user, HttpServletRequest request) {
-        Token token = tokenService.createToken(user,15);
+        Token token = tokenService.createToken(user, 15);
         String resetUrl = passwordService.createResetUrl(request, token);
         String subject = environment.getProperty("subject.password.message");
         String text = environment.getProperty("reset.password.message");
         text = String.format(text, "Maksim", "Stelmakh", resetUrl,
-            javaMailSender.getUsername());
+                javaMailSender.getUsername());
         sendSimpleMessage(user.getEmail(), subject, text);
+    }
 
-
+    private void sendSimpleMessage(String to, String subject, String text) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(to);
+        message.setSubject(subject);
+        message.setText(text);
+        emailSender.send(message);
     }
 }
