+++ src/test/java/com/epam/test_generator/controllers/SuitControllerSecurityTest.java
@@ -7,10 +7,9 @@ import com.epam.test_generator.dao.interfaces.UserDAO;
 import com.epam.test_generator.dto.LoginUserDTO;
 import com.epam.test_generator.entities.Role;
 import com.epam.test_generator.entities.User;
-import com.epam.test_generator.services.TokenService;
+import com.epam.test_generator.services.LoginService;
 import com.epam.test_generator.services.UserService;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
@@ -61,7 +60,7 @@ public class SuitControllerSecurityTest {
     private User validUser;
 
     @Autowired
-    private TokenService tokenService;
+    private LoginService loginService;
 
     @Mock
     private UserService userService;
@@ -96,7 +95,7 @@ public class SuitControllerSecurityTest {
         when(validUser.getRole()).thenReturn(new Role("GUEST"));
         when(validUser.isLocked()).thenReturn(false);
 
-        ReflectionTestUtils.setField(tokenService, "userService", userService);
+        ReflectionTestUtils.setField(loginService, "userService", userService);
     }
 
     @Test
@@ -105,7 +104,7 @@ public class SuitControllerSecurityTest {
         when(userService.getUserByEmail(anyString())).thenReturn(validUser);
         when(userService.isSamePasswords(anyString(), anyString())).thenReturn(true);
 
-        String token = "Bearer " + tokenService.getToken(loginUserDTO);
+        String token = "Bearer " + loginService.getLoginJWTToken(loginUserDTO);
 
         mvc.perform(get("/suits").header("Authorization", token).contentType("application/json"))
             .andDo(print())
@@ -118,7 +117,7 @@ public class SuitControllerSecurityTest {
         when(userService.getUserByEmail(anyString())).thenReturn(invalidUser);
         when(userService.isSamePasswords(anyString(), anyString())).thenReturn(true);
 
-        String token = "Bearer " + tokenService.getToken(loginUserDTO);
+        String token = "Bearer " + loginService.getLoginJWTToken(loginUserDTO);
 
         mvc.perform(get("/suits").header("Authorization", token).contentType("application/json"))
             .andDo(print())
@@ -142,7 +141,7 @@ public class SuitControllerSecurityTest {
         when(userService.getUserByEmail(anyString())).thenReturn(validUser);
         when(userService.isSamePasswords(anyString(), anyString())).thenReturn(true);
 
-        String token = "Bearer " + tokenService.getToken(loginUserDTO) + "something invalid";
+        String token = "Bearer " + loginService.getLoginJWTToken(loginUserDTO) + "something invalid";
 
         mvc.perform(get("/suits").header("Authorization", token).contentType("application/json"))
             .andDo(print())
@@ -155,7 +154,7 @@ public class SuitControllerSecurityTest {
         when(userService.getUserByEmail(anyString())).thenReturn(validUser);
         when(userService.isSamePasswords(anyString(), anyString())).thenReturn(true);
 
-        String token = "Bearer " + tokenService.getToken(loginUserDTO);
+        String token = "Bearer " + loginService.getLoginJWTToken(loginUserDTO);
 
         when(validUser.isLocked()).thenReturn(true);
